refactor admin tabs and handle loading/auth
